+++ test/languages/ini/comment.js
@@ -1,0 +1,12 @@
+exports.input = `
+;
+; foobar
+`;
+
+exports.expected = [
+    ["comment", ";"],
+    "\n",
+    ["comment", "; foobar"]
+];
+
+exports.comment = "Checks for comments";